PRESS0-2083 skip for non bluehost brand
